refactor(plugin-link)!: rename resolvers and fetchers
