+++ .yarn/versions/4df7cbc6.yml
@@ -1,0 +1,6 @@
+releases:
+  "@yarnpkg/plugin-link": major
+
+declined:
+  - "@yarnpkg/cli"
+  - "@yarnpkg/core"
