+++ packages/plugin-link/sources/PortalResolver.ts
@@ -1,0 +1,82 @@
+import {Resolver, ResolveOptions, MinimalResolveOptions} from '@yarnpkg/core';
+import {Descriptor, Locator, Manifest, Package}          from '@yarnpkg/core';
+import {LinkType}                                        from '@yarnpkg/core';
+import {miscUtils, structUtils}                          from '@yarnpkg/core';
+import {npath}                                           from '@yarnpkg/fslib';
+
+import {PORTAL_PROTOCOL}                                 from './constants';
+
+export class PortalResolver implements Resolver {
+  supportsDescriptor(descriptor: Descriptor, opts: MinimalResolveOptions) {
+    if (!descriptor.range.startsWith(PORTAL_PROTOCOL))
+      return false;
+
+    return true;
+  }
+
+  supportsLocator(locator: Locator, opts: MinimalResolveOptions) {
+    if (!locator.reference.startsWith(PORTAL_PROTOCOL))
+      return false;
+
+    return true;
+  }
+
+  shouldPersistResolution(locator: Locator, opts: MinimalResolveOptions) {
+    return false;
+  }
+
+  bindDescriptor(descriptor: Descriptor, fromLocator: Locator, opts: MinimalResolveOptions) {
+    return structUtils.bindDescriptor(descriptor, {
+      locator: structUtils.stringifyLocator(fromLocator),
+    });
+  }
+
+  getResolutionDependencies(descriptor: Descriptor, opts: MinimalResolveOptions) {
+    return {};
+  }
+
+  async getCandidates(descriptor: Descriptor, dependencies: unknown, opts: ResolveOptions) {
+    const path = descriptor.range.slice(PORTAL_PROTOCOL.length);
+
+    return [structUtils.makeLocator(descriptor, `${PORTAL_PROTOCOL}${npath.toPortablePath(path)}`)];
+  }
+
+  async getSatisfying(descriptor: Descriptor, dependencies: Record<string, Package>, locators: Array<Locator>, opts: ResolveOptions) {
+    const [locator] = await this.getCandidates(descriptor, dependencies, opts);
+
+    return {
+      locators: locators.filter(candidate => candidate.locatorHash === locator.locatorHash),
+      sorted: false,
+    };
+  }
+
+  async resolve(locator: Locator, opts: ResolveOptions): Promise<Package> {
+    if (!opts.fetchOptions)
+      throw new Error(`Assertion failed: This resolver cannot be used unless a fetcher is configured`);
+
+    const packageFetch = await opts.fetchOptions.fetcher.fetch(locator, opts.fetchOptions);
+
+    const manifest = await miscUtils.releaseAfterUseAsync(async () => {
+      return await Manifest.find(packageFetch.prefixPath, {baseFs: packageFetch.packageFs});
+    }, packageFetch.releaseFs);
+
+    return {
+      ...locator,
+
+      version: manifest.version || `0.0.0`,
+
+      languageName: manifest.languageName || opts.project.configuration.get(`defaultLanguageName`),
+      linkType: LinkType.SOFT,
+
+      conditions: manifest.getConditions(),
+
+      dependencies: opts.project.configuration.normalizeDependencyMap(manifest.dependencies),
+      peerDependencies: manifest.peerDependencies,
+
+      dependenciesMeta: manifest.dependenciesMeta,
+      peerDependenciesMeta: manifest.peerDependenciesMeta,
+
+      bin: manifest.bin,
+    };
+  }
+}
